WIP: don't allow Ipoll before Istackoffset
Failing test: tests/weak-ephe-final/weaklifetime (consistently failing
on Hephaestus - looks repeatable)

The end of the OCaml portion of Weak.create inserts an Ipoll before the
Istackoffset following the Iextcall to caml_weak_create. It's not at
present clear whether this is not a safe point or (more likely) the
frame_descriptor is not correctly indicating that %rax is live.

This repeatably results in the ephemeron not being marked and eventually
the page is freed which manifested (consistently) as the ephemeron
changing length!